refactor: 보드 상태를 변경하는 로직 훅으로 분리
